Terprint.Python/azureDataLake/requirements.txt
azure-storage-file-datalake>=12.8.0
azure-identity>=1.12.0
azure-mgmt-authorization>=3.0.0
azure-mgmt-storage>=20.0.0
azure-mgmt-resource>=21.0.0
azure-graphrbac>=0.61.1